% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divid.R
\name{/}
\alias{/}
\title{minus}
\usage{
e1 / e2
}
\arguments{
\item{e1}{Left operand}

\item{e2}{Right operand}
}
\value{
Result of e1 + e2
}
\description{
Override the + operator to perform subtraction
}
\author{
Joy Ding
}
